requirements.txt
deepagents==0.2.5
tavily-python==0.7.12
python-dotenv==1.2.1
langchain-google-genai==3.0.1
jupyterlab==4.4.10
ipykernel==7.1.0
google-ai-generativelanguage==0.9.0
google-auth==2.43.0
google-api-core==2.28.1